# api/openapi.yaml
openapi: 3.0.3
info:
  title: Coding Agent CLI REST API
  description: |
    REST API for the Coding Agent CLI security scanning platform.
    
    This API provides programmatic access to security scanning, findings management,
    policy validation, analytics, and webhook configuration.
    
    ## Features
    - Trigger and manage security scans
    - Query and filter security findings
    - Validate security policies
    - Access analytics and trends
    - Configure webhook notifications
    - Import findings from external tools
    
    ## Base URL
    The default base URL is `http://localhost:8080/api/v1`
    
    ## Rate Limiting
    - Unauthenticated: 100 requests/minute
    - Authenticated: 1000 requests/minute
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
  version: 1.2.0
  contact:
    name: Coding Agent CLI
    url: https://github.com/coding-agent/cli
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: http://localhost:{port}/api/v1
    description: Custom port server
    variables:
      port:
        default: '8080'
        description: Server port
...
paths:
...
  /analytics/trends:
    get:
      tags:
        - Analytics
      summary: Get finding trends
      description: |
        Retrieve historical trend data for security findings.
        Shows how finding counts by severity have changed over time.
      operationId: getTrends
      parameters:
        - name: start
          in: query
          description: Start date (ISO 8601)
          schema:
            type: string
            format: date-time
          example: '2026-01-01T00:00:00Z'
        - name: end
          in: query
          description: End date (ISO 8601)
          schema:
            type: string
            format: date-time
          example: '2026-03-01T00:00:00Z'
        - name: interval
          in: query
          description: Time interval for grouping
          schema:
            type: string
            enum: [day, week, month]
            default: week
          example: week
      responses:
        '200':
          description: Trend data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendsResponse'
              example:
                dates:
                  - '2026-01-01'
                  - '2026-01-08'
                  - '2026-01-15'
                critical: [5, 3, 2]
                high: [12, 10, 8]
                medium: [25, 22, 20]
                low: [15, 14, 12]
        '501':
          description: Analytics not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
components:
...
  schemas:
...
    TrendsResponse:
      type: object
      required:
        - dates
        - critical
        - high
        - medium
        - low
      properties:
        dates:
          type: array
          description: Date labels for trend data
          items:
            type: string
            format: date
          example: ['2026-01-01', '2026-01-08', '2026-01-15']
        critical:
          type: array
          description: Critical finding counts
          items:
            type: integer
          example: [5, 3, 2]
        high:
          type: array
          description: High severity finding counts
          items:
            type: integer
          example: [12, 10, 8]
        medium:
          type: array
          description: Medium severity finding counts
          items:
            type: integer
          example: [25, 22, 20]
        low:
          type: array
          description: Low severity finding counts
          items:
            type: integer
          example: [15, 14, 12]
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          enum:
            - INVALID_REQUEST
            - NOT_FOUND
            - UNAUTHORIZED
            - INTERNAL_ERROR
            - SCANNER_FAILED
            - LLM_FAILED
          example: INVALID_REQUEST
        message:
          type: string
          description: Human-readable error message
          example: Path is required
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          example:
            field: path
            constraint: required
...